+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Fixes
 - Fixed issue when slicing multiple objects and outputting bincode overlapping and conflicting 
 - Removed all unwraps. Converted some to errors and some to expect. deny(clippy::unwrap_used) is added to prevent unwrap being added in future
+- Added checks for models and moves to make sure they don't go out of bounds
 
 ## [0.3.0]
 ### New Features